product-landing
For Assignment Week #6....boy this was a mess XD
Code number 1 was my best attempt at following the instructions given.   But when it didn't work, I made code number 2.
Code number 2 doesn't follow the instructions to a T, but it, well, actually works.  Not my best, but, better than nothing!